Pointer Arrays and Function Pointers in C
This C program demonstrates two key pointer concepts:
- Pointer to an Array:
The function fill accepts a pointer to an array of 10 integers and fills it with the squares of indices (i * i). Although this function is defined, it is not invoked in main, serving as a reference for how to pass arrays to functions using pointers.
- Array of Pointers to Integers:
In main, three integers (a, b, and c) are declared and their addresses are stored in an array of pointers arr[3]. A loop is used to dereference and print the values pointed to by each element of the array.